seems to be working okay
Got a handle on how to use asyncMap.  the callback passed is only called
if there is an error condition, or if the list of items is done.

Can probably modify the second list to use asyncMap as well, but that would
take tweaking to make sure the queue of calls doesn't get huge

 modified:   lib/detector_data.js